Move importing implementation methods to GVRImporter. GVRContext APIs are preserved.
