+++ GVRf/Framework/src/org/gearvrf/GVRContext.java
@@ -26,21 +26,11 @@ import org.gearvrf.GVRAndroidResource.BitmapTextureCallback;
 import org.gearvrf.GVRAndroidResource.CompressedTextureCallback;
 import org.gearvrf.GVRAndroidResource.MeshCallback;
 import org.gearvrf.GVRAndroidResource.TextureCallback;
-import org.gearvrf.GVRMaterial.GVRShaderType;
 import org.gearvrf.animation.GVRAnimation;
 import org.gearvrf.animation.GVRAnimationEngine;
 import org.gearvrf.asynchronous.GVRAsynchronousResourceLoader;
 import org.gearvrf.asynchronous.GVRCompressedTexture;
 import org.gearvrf.asynchronous.GVRCompressedTextureLoader;
-import org.gearvrf.jassimp.AiColor;
-import org.gearvrf.jassimp.AiMaterial;
-import org.gearvrf.jassimp.AiNode;
-import org.gearvrf.jassimp.AiScene;
-import org.gearvrf.jassimp.AiTextureType;
-import org.gearvrf.jassimp.GVROldWrapperProvider;
-import org.gearvrf.jassimp2.GVRJassimpAdapter;
-import org.gearvrf.jassimp2.GVRJassimpSceneObject;
-import org.gearvrf.jassimp2.Jassimp;
 import org.gearvrf.periodic.GVRPeriodicEngine;
 import org.gearvrf.utility.Log;
 import org.gearvrf.utility.ResourceCache;
@@ -490,8 +480,6 @@ public abstract class GVRContext {
                 GVRImportSettings.getRecommendedSettings());
     }
 
-    GVROldWrapperProvider wrapperProvider = new GVROldWrapperProvider();
-
     /**
      * Simple, high-level method to load a scene as {@link GVRSceneObject} from
      * 3D model.
@@ -513,23 +501,7 @@ public abstract class GVRContext {
      */
     public GVRSceneObject getAssimpModel(String assetRelativeFilename,
             EnumSet<GVRImportSettings> settings) throws IOException {
-        
-        GVRAssimpImporter assimpImporter = GVRImporter.readFileFromResources(
-                this, new GVRAndroidResource(this, assetRelativeFilename),
-                settings);
-        
-        GVRSceneObject wholeSceneObject = new GVRSceneObject(this);
-
-        AiScene assimpScene = assimpImporter.getAssimpScene();
-
-        AiNode rootNode = assimpScene.getSceneRoot(wrapperProvider);
-
-        // Recurse through the entire hierarchy to attache all the meshes as
-        // Scene Object
-        recurseAssimpNodes(assimpImporter, assetRelativeFilename, wholeSceneObject,
-                rootNode, wrapperProvider) ;
-
-        return wholeSceneObject;
+        return GVRImporter.getAssimpModel(this, assetRelativeFilename, settings);
     }
 
     public GVRSceneObject loadJassimpModelFromSD(String externalFile) throws IOException {
@@ -537,12 +509,7 @@ public abstract class GVRContext {
     }
 
     public GVRSceneObject loadJassimpModelFromSD(String externalFile, EnumSet<GVRImportSettings> settings) throws IOException {
-        Jassimp.setWrapperProvider(GVRJassimpAdapter.sWrapperProvider);
-        org.gearvrf.jassimp2.AiScene assimpScene = Jassimp.importFile(externalFile,
-            GVRJassimpAdapter.get().toJassimpSettings(settings));
-        if (assimpScene == null)
-            return null;
-        return new GVRJassimpSceneObject(this, assimpScene);
+        return GVRImporter.loadJassimpModelFromSD(this, externalFile, settings);
     }
 
     public GVRSceneObject loadJassimpModel(String assetFile) throws IOException {
@@ -550,213 +517,7 @@ public abstract class GVRContext {
     }
 
     public GVRSceneObject loadJassimpModel(String assetFile, EnumSet<GVRImportSettings> settings) throws IOException {
-        Jassimp.setWrapperProvider(GVRJassimpAdapter.sWrapperProvider);
-        org.gearvrf.jassimp2.AiScene assimpScene = Jassimp.importAssetFile(assetFile,
-            GVRJassimpAdapter.get().toJassimpSettings(settings), mContext.getAssets());
-        if (assimpScene == null)
-            return null;
-        return new GVRJassimpSceneObject(this, assimpScene);
-    }
-
-    /**
-     * Helper method to recurse through all the assimp nodes and get all their
-     * meshes that can be used to create {@link GVRSceneObject} to be attached
-     * to the set of complete scene objects for the assimp model.
-     * 
-     * @param assetRelativeFilename
-     *            A filename, relative to the {@code assets} directory. The file
-     *            can be in a sub-directory of the {@code assets} directory:
-     *            {@code "foo/bar.png"} will open the file
-     *            {@code assets/foo/bar.png}
-     * 
-     * @param wholeSceneObject
-     *            A reference of the {@link GVRSceneObject}, to which all other
-     *            scene objects are attached.
-     * 
-     * @param node
-     *            A reference to the AiNode for which we want to recurse all its
-     *            children and meshes.
-     * 
-     * @param wrapperProvider
-     *            AiWrapperProvider for unwrapping Jassimp properties.
-     * 
-     */
-
-    @SuppressWarnings("resource")
-    private void recurseAssimpNodes(
-            GVRAssimpImporter assimpImporter,
-            String assetRelativeFilename,
-            GVRSceneObject parentSceneObject,
-            AiNode node,
-            GVROldWrapperProvider wrapperProvider) {
-        try {
-            GVRSceneObject newParentSceneObject = new GVRSceneObject(this);
-
-            if (node.getNumMeshes() == 0) {
-                parentSceneObject.addChildObject(newParentSceneObject);
-                parentSceneObject = newParentSceneObject;
-            } else if (node.getNumMeshes() == 1) {
-                // add the scene object to the scene graph
-                GVRSceneObject sceneObject = createSceneObject(
-                        assimpImporter, assetRelativeFilename, node, 0, wrapperProvider);
-                parentSceneObject.addChildObject(sceneObject);
-                parentSceneObject = sceneObject;
-            } else {
-                for (int i = 0; i < node.getNumMeshes(); i++) {
-                    GVRSceneObject sceneObject = createSceneObject(
-                            assimpImporter, assetRelativeFilename, node, i, wrapperProvider);
-                    newParentSceneObject.addChildObject(sceneObject);
-                }
-                parentSceneObject.addChildObject(newParentSceneObject);
-                parentSceneObject = newParentSceneObject;
-            }
-
-            if (node.getTransform(wrapperProvider) != null) {
-                parentSceneObject.getTransform().setModelMatrix(
-                		GVROldWrapperProvider.transpose(node.getTransform(wrapperProvider).toByteBuffer()));
-            }
-            
-            for (int i = 0; i < node.getNumChildren(); i++) {
-                this.recurseAssimpNodes(assimpImporter, assetRelativeFilename,
-                        parentSceneObject, node.getChildren().get(i),
-                        wrapperProvider);
-            }
-        } catch (Exception e) {
-            // Error while recursing the Scene Graph
-            e.printStackTrace();
-        }
-    }
-
-    private float[] transpose(float[] modelMatrix){
-        float[] transposed = new float[16];
-        transposed[0] = modelMatrix[0];
-        transposed[4] = modelMatrix[1];
-        transposed[8] = modelMatrix[2];
-        transposed[12] = modelMatrix[3];
-        transposed[1] = modelMatrix[4];
-        transposed[5] = modelMatrix[5];
-        transposed[9] = modelMatrix[6];
-        transposed[13] = modelMatrix[7];
-        transposed[2] = modelMatrix[8];
-        transposed[6] = modelMatrix[9];
-        transposed[10] = modelMatrix[10];
-        transposed[14] = modelMatrix[11];
-        transposed[3] = modelMatrix[12];
-        transposed[7] = modelMatrix[13];
-        transposed[11] = modelMatrix[14];
-        transposed[15] = modelMatrix[15];
-        return transposed;
-    }
-
-    /**
-     * Helper method to create a new {@link GVRSceneObject} with the mesh at the
-     * index {@link index} of the node mesh array with a color or texture
-     * material.
-     * 
-     * @param assetRelativeFilename
-     *            A filename, relative to the {@code assets} directory. The file
-     *            can be in a sub-directory of the {@code assets} directory:
-     *            {@code "foo/bar.png"} will open the file
-     *            {@code assets/foo/bar.png}
-     * 
-     * @param node
-     *            A reference to the AiNode for which we want to recurse all its
-     *            children and meshes.
-     * 
-     * @param index
-     *            The index of the mesh in the array of meshes for that node.
-     * 
-     * @param wrapperProvider
-     *            AiWrapperProvider for unwrapping Jassimp properties.
-     * 
-     * @return The new {@link GVRSceneObject} with the mesh at the index
-     *         {@link index} for the node {@link node}
-     * 
-     * @throws IOException
-     *             File does not exist or cannot be read
-     */
-    private GVRSceneObject createSceneObject(
-            GVRAssimpImporter assimpImporter,
-            String assetRelativeFilename,
-            AiNode node,
-            int index,
-            GVROldWrapperProvider wrapperProvider)
-            throws IOException {
-        
-        FutureWrapper<GVRMesh> futureMesh = new FutureWrapper<GVRMesh>(
-                this.getNodeMesh(assimpImporter, node.getName(), index));
-
-        AiMaterial material = this.getMeshMaterial(assimpImporter, node.getName(), index);
-        
-        final GVRMaterial meshMaterial = new GVRMaterial(this,
-                GVRShaderType.Assimp.ID);
-     
-        /* Feature set */
-        int assimpFeatureSet = 0x00000000;
-        
-        /* Diffuse color */
-        AiColor diffuseColor = material.getDiffuseColor(wrapperProvider);
-        meshMaterial.setDiffuseColor(diffuseColor.getRed(),
-                diffuseColor.getGreen(), diffuseColor.getBlue(),
-                diffuseColor.getAlpha());
-
-        /* Specular color */
-        AiColor specularColor = material.getSpecularColor(wrapperProvider);
-        meshMaterial.setSpecularColor(specularColor.getRed(),
-                specularColor.getGreen(), specularColor.getBlue(),
-                specularColor.getAlpha());
-
-        /* Ambient color */
-        AiColor ambientColor = material.getAmbientColor(wrapperProvider);
-        meshMaterial.setAmbientColor(ambientColor.getRed(),
-                ambientColor.getGreen(), ambientColor.getBlue(),
-                ambientColor.getAlpha());
-
-        /* Emissive color */
-        AiColor emissiveColor = material.getEmissiveColor(wrapperProvider);
-        meshMaterial.setVec4("emissive_color", emissiveColor.getRed(),
-                emissiveColor.getGreen(), emissiveColor.getBlue(),
-                emissiveColor.getAlpha());
-
-        /* Opacity */
-        float opacity = material.getOpacity();
-        meshMaterial.setOpacity(opacity);
-
-        /* Diffuse Texture */
-        final String texDiffuseFileName = material.getTextureFile(
-                AiTextureType.DIFFUSE, 0);
-        if (texDiffuseFileName != null && !texDiffuseFileName.isEmpty()) {
-            loadTexture(new TextureCallback() {
-                @Override
-                public void loaded(GVRTexture texture, GVRAndroidResource ignored) {
-                    meshMaterial.setMainTexture(texture);
-                    final int features = GVRShaderType.Assimp.setBit(
-                            meshMaterial.getShaderFeatureSet(),
-                            GVRShaderType.Assimp.AS_DIFFUSE_TEXTURE);
-                    meshMaterial.setShaderFeatureSet(features);
-                }
-                @Override
-                public void failed(Throwable t, GVRAndroidResource androidResource) {
-                    Log.e(TAG, "Error loading diffuse texture %s; exception: %s",
-                        texDiffuseFileName, t.getMessage());
-                }
-                @Override
-                public boolean stillWanted(GVRAndroidResource androidResource) {
-                    return true;
-                }
-            }, new GVRAndroidResource(this, texDiffuseFileName));
-        }
- 
-        /* Apply feature set to the material */
-        meshMaterial.setShaderFeatureSet(assimpFeatureSet);
-
-        GVRSceneObject sceneObject = new GVRSceneObject(this);
-        sceneObject.setName(node.getName());
-        GVRRenderData sceneObjectRenderData = new GVRRenderData(this);
-        sceneObjectRenderData.setMesh(futureMesh);
-        sceneObjectRenderData.setMaterial(meshMaterial);
-        sceneObject.attachRenderData(sceneObjectRenderData);
-        return sceneObject;
+        return GVRImporter.loadJassimpModel(this, assetFile, settings);
     }
 
     /**
@@ -771,16 +532,6 @@ public abstract class GVRContext {
     }
 
     /**
-     * Retrieves the material for the mesh of the given node..
-     * 
-     * @return The material, encapsulated as a {@link AiMaterial}.
-     */
-    public AiMaterial getMeshMaterial(GVRAssimpImporter assimpImporter,
-            String nodeName, int meshIndex) {
-        return assimpImporter.getMeshMaterial(nodeName, meshIndex);
-    }
-
-    /**
      * Creates a quad consisting of two triangles, with the specified width and
      * height.
      * 
@@ -790,6 +541,7 @@ public abstract class GVRContext {
      *            the quad's height
      * @return A 2D, rectangular mesh with four vertices and two triangles
      */
+    @SuppressWarnings("deprecation")
     public GVRMesh createQuad(float width, float height) {
         GVRMesh mesh = new GVRMesh(this);
